feat!: rename utils.py to menu.py and create a new utils.py with some functions
