+++ src/structures/save_games_info.txt
@@ -1,2 +1,3 @@
-1
+2
 zuca
+tuna
